requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asyncpg==0.29.0
click==8.1.7
dnspython==2.7.0
fastapi==0.115.2
greenlet==3.1.1
h11==0.14.0
idna==3.10
Jinja2==3.1.4
MarkupSafe==3.0.1
psycopg-binary==3.1.19
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic_core==2.23.4
pyIsEmail==2.0.1
python-dotenv==1.0.1
python-multipart==0.0.12
setuptools==75.1.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette~=0.40.0
typing_extensions==4.12.2
uvicorn==0.31.1
passlib~=1.7.4